Inline simple_bitmask methods; drop dll export to fix C4273
simple_bitmask was 'class DISKANN_DLLEXPORT', but consumers that compile
label_bitmask.cpp directly (e.g. AdsSnr's ANNTestTool, which defines neither
_WINDLL nor DISKANN_STATIC_LIB) then define its members in a dllimport context,
triggering C4273 'inconsistent dll linkage' (a build error under CloudBuild
warning-as-error/auto-baseline).

Make all simple_bitmask methods inline (header-only) and remove the class-level
DISKANN_DLLEXPORT; label_bitmask.cpp now has no out-of-line definitions. This
resolves the odr-use from ColorInfoVector's inline constructor (color_info.h)
locally in every consumer -- DLL, static lib, and direct-compile -- with no
import/export, eliminating C4273.
